+++ app/components/layout/header-icon-base.js
@@ -9,11 +9,11 @@ import { vars } from '../../styles/styles';
 export default class HeaderIconBase extends SafeComponent {
     icon = '';
     action = () => { };
-
+    disabled = false;
     renderThrow() {
         return (
-            <View style={[{ flex: 0 }, this.style]}>
-                <TouchableOpacity onPress={this.action}>
+            <View style={[{ flex: 0, opacity: this.disabled ? 0.5 : 1 }, this.style]}>
+                <TouchableOpacity onPress={this.disabled ? null : this.action} activeOpacity={this.disabled ? 0.5 : 1}>
                     <View style={[{
                         alignItems: 'center',
                         backgroundColor: 'transparent',