+++ app/components/layout/loading-screen.js
@@ -47,6 +47,7 @@ export default class LoadingScreen extends Component {
     async componentDidMount() {
         try {
             await loginState.load();
+            if (!loginState.loaded) throw new Error('error logging in after return');
             this.goToNextStep();
             await promiseWhen(() => socket.authenticated);
             this.goToNextStep();
@@ -55,10 +56,11 @@ export default class LoadingScreen extends Component {
             await promiseWhen(() => routerMain.fileStateLoaded);
             this.goToNextStep();
             await promiseWhen(() => routerMain.contactStateLoaded);
-            if (!loginState.loaded) routerApp.routes.loginStart.transition();
         } catch (e) {
             console.log('loading-screen.js: loading screen error');
+            if (!loginState.loaded) routerApp.routes.loginStart.transition();
             console.error(e);
+            return;
         }
         this.fadeInOut();
         this.growIcon();
@@ -141,7 +143,7 @@ export default class LoadingScreen extends Component {
         };
         this.iconState = this.currentState;
         return (
-            <View style={{ flexDirection: 'row', height: 48, justifyContent: 'center', alignItems: 'center' }}>
+            <View key={name} style={{ flexDirection: 'row', height: 48, justifyContent: 'center', alignItems: 'center' }}>
                 <Animated.Image
                     key={`${name}Icon`}
                     source={this.iconState[`${name}`].icon}